pyproject.toml
[project]
name = "ee-inventory-management"
version = "0.1.0"
description = "E-Bike Inventory Management System"
requires-python = ">=3.11"
dependencies = [
    "flask==3.1.2",
    "flask-cors==5.0.1",
    "google-genai>=1.62.0",
    "python-barcode[images]==0.15.1",
    "reportlab==4.2.5",
    "pydantic==2.10.6",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "ruff>=0.9",
    "responses>=0.25",
    "types-requests>=2.32",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "database", "services"]

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
    "S",   # flake8-bandit (security)
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]  # allow assert in tests
"main.py" = ["T20"]    # allow print in CLI

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
]
strict_markers = true

[tool.mypy]
strict = true
python_version = "3.11"

[[tool.mypy.overrides]]
module = [
    "barcode.*",
    "reportlab.*",
    "google.genai.*",
    "flask_cors",
]
ignore_missing_imports = true
